requirements.txt
# Dashboard
streamlit>=1.28.0

# Data & Analysis
pandas>=2.0.0
plotly>=5.18.0

# Scraping
apify-client>=1.6.0
requests>=2.31.0

# NLP & Text Analysis
textblob>=0.17.1

# Visualization
wordcloud>=1.9.2
matplotlib>=3.8.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Date handling
python-dateutil>=2.8.2
